Matrix is now created using Int64

+++ highs/ipm/hipo/factorhighs/FactorHiGHS.cpp
@@ -34,15 +34,15 @@ void FHsolver::setRegularisation(double reg_p, double reg_d) {
   regul_.dual = reg_d;
 }
 
-Int FHsolver::analyse(Symbolic& S, const std::vector<Int>& rows,
-                      const std::vector<Int>& ptr,
+Int FHsolver::analyse(Symbolic& S, const std::vector<Int64>& rows,
+                      const std::vector<Int64>& ptr,
                       const std::vector<Int>& signs) {
   Analyse an_obj(rows, ptr, signs, nb_, log_, data_);
   return an_obj.run(S);
 }
 
-Int FHsolver::factorise(const Symbolic& S, const std::vector<Int>& rows,
-                        const std::vector<Int>& ptr,
+Int FHsolver::factorise(const Symbolic& S, const std::vector<Int64>& rows,
+                        const std::vector<Int64>& ptr,
                         const std::vector<double>& vals) {
   Factorise fact_obj(S, rows, ptr, vals, regul_, log_, data_, sn_columns_);
   return fact_obj.run(N_);
